Move OAuth logic to separate file
